♻️ Dividing components into smaller parts, small fixes
